product_service/requirements.txt
aniso8601==9.0.1
anyio==3.6.2
apt-xapian-index==0.49
arrow==1.2.3
attrs==21.2.0
binaryornot==0.4.4
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cookiecutter==2.1.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fastapi==0.91.0
Flask==2.2.2
flask-restplus==0.13.0
Flask-SQLAlchemy==3.0.3
fuse-python==1.0.2
gpg===1.16.0-unknown
greenlet==2.0.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
pexpect==4.8.0
Pillow==9.0.1
psutil==5.9.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.4
PyGObject==3.42.1
PyJWT==2.3.0
pylibacl==0.6.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-slugify==8.0.0
pytz==2022.7.1
pyxattr==0.7.2
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.24.0
systemd-python==234
terminator==2.1.1
text-unidecode==1.3
texttable==1.6.4
tornado==6.1
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==2.2.2
xkit==0.0.0
zipp==1.0.0
